requirements.txt
fastapi
uvicorn
pydantic
fpdf2
python-multipart